complete port

+++ wekit-zygisk/native/src/companion.rs
@@ -1,9 +1,11 @@
-// companion — Zygisk root 进程请求处理
+// companion — Zygisk root process request handler.
 //
-// 检查注入 allow-list，处理启用查询或协商 Telegram worker socket。
-// Telegram worker 双 fork 由 init 收养，每个连接先检查目标是否仍启用，
-// 再处理 DISCOVER（查找 /data/user/{uid} 下已知 Telegram 包）
-// 或 COPY_DATABASE（发送 cache4.db + wal + shm 文件内容）。
+// Checks the injection allow-list, handles enable/disable queries,
+// and negotiates a Telegram worker socket via double-fork.
+#![allow(clippy::unnecessary_cast)]
+// The worker is adopted by init; each connection checks the target
+// is still enabled, then handles DISCOVER (/data/user/{uid} known
+// Telegram packages) or COPY_DATABASE (cache4.db + wal + shm).
 
 use crate::protocol::*;
 use crate::{loge, logi, logw};
@@ -50,7 +52,7 @@ fn is_enabled_target(uid: i32, process_name: &str) -> bool {
     false
 }
 
-// ── Telegram 包识别 ────────────────────────────────────────────────────────────
+// ── Telegram package identification ────────────────────────────────────────────
 
 static KNOWN_TELEGRAM_PACKAGES: &[&str] = &[
     "org.telegram.messenger",
@@ -99,9 +101,9 @@ fn is_known_telegram_package(pkg: &str) -> bool {
     KNOWN_TELEGRAM_PACKAGES.contains(&pkg)
 }
 
-// ── Telegram 数据库路径 ────────────────────────────────────────────────────────
+// ── Telegram database path ────────────────────────────────────────────────────────
 
-/// 读取 shared_prefs/userconfing.xml 获取 selectedAccount，对应 C++ read_selected_account。
+/// Parse shared_prefs/userconfing.xml for selectedAccount.
 fn read_selected_account(app_dir: &str) -> i32 {
     let config = format!("{app_dir}/shared_prefs/userconfing.xml");
     let content = match fs::read_to_string(&config) {
@@ -124,8 +126,8 @@ fn read_selected_account(app_dir: &str) -> i32 {
     0
 }
 
-/// 返回 Telegram 数据库文件路径（cache4.db），不存在则返回空字符串。
-/// 路径：/data/user/{user_id}/{pkg}/files/[account{N}/]cache4.db
+/// Return the Telegram database path (cache4.db), or empty if not found.
+/// Path: /data/user/{user_id}/{pkg}/files/[account{N}/]cache4.db
 fn telegram_database_path(user_id: i32, pkg: &str) -> String {
     if user_id < 0 || !is_valid_telegram_package(pkg) {
         return String::new();
@@ -143,7 +145,7 @@ fn telegram_database_path(user_id: i32, pkg: &str) -> String {
     };
     let mut st: libc::stat = unsafe { std::mem::zeroed() };
     if unsafe { libc::stat(cdb.as_ptr(), &mut st) } == 0
-        && (st.st_mode & libc::S_IFMT) == libc::S_IFREG
+        && (st.st_mode & libc::S_IFMT as u32) == libc::S_IFREG as u32
     {
         db
     } else {
@@ -151,7 +153,7 @@ fn telegram_database_path(user_id: i32, pkg: &str) -> String {
     }
 }
 
-/// 扫描 /data/user/{user_id}/ 下已知的 Telegram 包，对应 C++ discover_telegram_packages。
+/// Scan /data/user/{user_id}/ for known Telegram packages with a valid database.
 fn discover_telegram_packages(user_id: i32) -> Vec<String> {
     let user_dir = format!("/data/user/{user_id}");
     let mut packages = Vec::new();
@@ -172,9 +174,9 @@ fn discover_telegram_packages(user_id: i32) -> Vec<String> {
     packages
 }
 
-// ── 文件发送 ──────────────────────────────────────────────────────────────────
+// ── File sending ──────────────────────────────────────────────────────────────────
 
-/// 发送文件内容：先写 u64 size，再发文件字节；required=false 时文件不存在发 0。
+/// Send file contents: write u64 size then file bytes; required=false sends 0 on ENOENT.
 fn send_file(sock: c_int, path: &str, required: bool) -> bool {
     let cpath = match CString::new(path) {
         Ok(s) => s,
@@ -194,7 +196,7 @@ fn send_file(sock: c_int, path: &str, required: bool) -> bool {
         );
         return false;
     }
-    if (st.st_mode & libc::S_IFMT) != libc::S_IFREG {
+    if (st.st_mode & libc::S_IFMT as u32) != libc::S_IFREG as u32 {
         loge!("Zygisk: companion: {} is not a regular file", path);
         return false;
     }
@@ -270,7 +272,7 @@ fn write_error_response(sock: c_int, msg: &str) {
 
 // ── Telegram worker ────────────────────────────────────────────────────────────
 
-/// 处理单个 telegram 请求，对应 C++ handle_telegram_request。
+/// Handle a single Telegram request from a connected client.
 fn handle_telegram_request(client: c_int, user_id: i32) -> bool {
     let op = match read_u8_from_fd(client) {
         Ok(v) => v,
@@ -321,7 +323,7 @@ fn handle_telegram_request(client: c_int, user_id: i32) -> bool {
         && send_file(client, &format!("{db}-shm"), false)
 }
 
-/// 后台 worker：循环 accept，每连接检查 is_enabled_target，对应 C++ telegram_worker。
+/// Background worker: loop accept, check is_enabled_target per connection, serve or reject.
 fn telegram_worker(server_fd: c_int, uid: i32, process_name: String) {
     let user_id = uid / APP_USER_RANGE;
     loop {
@@ -350,7 +352,7 @@ fn telegram_worker(server_fd: c_int, uid: i32, process_name: String) {
     unsafe { libc::close(server_fd) };
 }
 
-// ── 请求头读取 ─────────────────────────────────────────────────────────────────
+// ── Request header ─────────────────────────────────────────────────────────────────
 
 struct RequestHeader {
     request_type: u8,
@@ -381,9 +383,9 @@ fn random_nonce() -> u32 {
     u32::from_ne_bytes(buf)
 }
 
-// ── 入口 ──────────────────────────────────────────────────────────────────────
+// ── Entry point ──────────────────────────────────────────────────────────────────────
 
-/// 由 zygisk_companion_entry 调用，对应 C++ companion_handler。
+/// Called from zygisk_companion_entry in lib.rs.
 pub fn handle(sock: c_int) {
     let header = match read_header(sock) {
         Some(h) => h,
@@ -416,7 +418,7 @@ pub fn handle(sock: c_int) {
         return;
     }
 
-    // 创建 abstract Unix socket
+    // Create abstract Unix socket
     let fd = unsafe { libc::socket(AF_UNIX, SOCK_STREAM | libc::SOCK_CLOEXEC, 0) };
     if fd < 0 {
         loge!(
@@ -447,7 +449,7 @@ pub fn handle(sock: c_int) {
         return;
     }
 
-    // 双 fork：中间子进程立即退出，孙子进程由 init 收养
+    // Double-fork: intermediate child exits, grandchild adopted by init
     let uid = header.uid;
     let process_name = header.process_name.clone();
     let mid = unsafe { libc::fork() };
@@ -461,19 +463,19 @@ pub fn handle(sock: c_int) {
         return;
     }
     if mid == 0 {
-        // 中间子进程
+        // Intermediate child
         let worker = unsafe { libc::fork() };
         if worker < 0 {
             unsafe { libc::_exit(1) };
         }
         if worker == 0 {
-            // 孙子进程（真正的 worker）
+            // Grandchild (real worker)
             telegram_worker(fd, uid, process_name);
             unsafe { libc::_exit(0) };
         }
-        unsafe { libc::_exit(0) }; // 中间进程立即退出
+        unsafe { libc::_exit(0) }; // intermediate exits immediately
     }
-    // 父进程：等中间子进程退出，关闭自己持有的 server_fd
+    // Parent: wait for intermediate child, close our copy of server_fd
     unsafe {
         let mut status = 0i32;
         libc::waitpid(mid, &mut status, 0);
